backend/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
PyJWT==2.10.1
stripe==12.5.1
Werkzeug==3.1.3
